test: gate firecracker fork tests behind linux build tag
The two TestForkFirecracker* tests invoke forkInstanceFromStoppedOrStandby
which needs a hypervisor VM starter — firecracker is linux-only, so these
fail on darwin with 'no VM starter for hypervisor type: firecracker'.

Split into _linux_test.go; leave the pure-helper TestInstallForkSharedMemFile_*
tests cross-platform.
